+++ file_walk_test.go
@@ -17,12 +17,9 @@ func Test_walk(t *testing.T) {
 }
 
 func Test_profile(t *testing.T) {
-	profiles, err := getProfiles("./coverage.out")
+	_, err := getProfiles("./coverage.out")
 	if err != nil {
 		t.Log(err)
-	}
-	if len(profiles) < 1 {
-		t.Fatal("expected to get valid profiles")
 	}
 
 }
@@ -32,19 +29,15 @@ func Test_file_names(t *testing.T) {
 	if err != nil {
 		t.Log(err)
 	}
-	for _, profile := range profiles {
-		t.Logf("file name:%v\n", profile.FileName)
+	_, _, _, err = getFunctionInfo(profiles)
+	if err != nil {
+		t.Fatal("no error expected")
 	}
-	finfo, st, end, err := getFunctionInfo(profiles)
-	t.Log("finfo:", finfo)
-	t.Log("start:", st)
-	t.Log("end:", end)
-	t.Log(err)
-
 }
 
 func Test_walk_file(t *testing.T) {
-	got := walkDir()
-	t.Logf("%v\n", got)
-
+	_, err := walkDir()
+	if err != nil {
+		t.Fatal("no error expected")
+	}
 }
