--- internal/testdata/test_coverage.out
@@ -1,5 +1,0 @@
-mode: atomic
-github.com/ehrktia/venn/intersection.go:5.48,9.32 4 3
-github.com/ehrktia/venn/intersection.go:14.2,14.10 1 3
-github.com/ehrktia/venn/intersection.go:9.32,10.32 1 3
-github.com/ehrktia/venn/intersection.go:10.32,12.4 1 1
